Making common text snippet commands, extending list
Common text snippets are very useful!  Just as we do for the coordinate
footnote, we should make sure they are commands that can be used.  They
should probably also exist in a style file, since it's a list of
commands, rather than as a template file that gets included in latex
(won't make a functional difference, but keeps the style of the template
in order, I think).

I made both of these changes, and updated the note and paper templates
to point to the correct places.

Also added a few additional snippets based on the current twiki doc at
https://twiki.cern.ch/twiki/bin/view/AtlasProtected/PubComCommonText

If these changes are accepted, then I'd suggest adding more snippets of
this sort from that twiki, in coordination with PMG and the luminosity
group in particular (the heaviest users).